+++ ui/components/Policies/Utils/PolicyUtils.tsx
@@ -1,6 +1,9 @@
-import React from "react";
-import { Chip } from "@material-ui/core";
+import * as React from "react";
 import styled from "styled-components";
+import ReactMarkdown from "react-markdown";
+import { Chip } from "@material-ui/core";
+import Flex from "../../Flex";
+import Text from "../../Text";
 
 export const parseValue = (parameter: {
   type?: string | undefined;
@@ -19,15 +22,45 @@ export const parseValue = (parameter: {
   }
 };
 
+export const SectionWrapper = ({ title, children }) => {
+  return (
+    <Flex column wide gap="8" data-testid="occurrences">
+      <Text bold color="neutral30">
+        {title}
+      </Text>
+      {children}
+    </Flex>
+  );
+};
+
 export const ChipWrap = styled(Chip)`
   &.MuiChip-root {
     color: ${(props) => props.theme.colors.black};
-    background-color: ${(props) => props.theme.colors.neutralGray};
+    background-color: ${(props) => props.theme.colors.neutral10};
     padding: 2px 4px;
     height: inherit;
     border-radius: 4px;
   }
   .MuiChip-label {
     padding: 0;
+  }
+`;
+
+export const Editor = styled(ReactMarkdown)`
+  width: calc(100% - 24px);
+  padding: ${(props) => props.theme.spacing.small};
+  overflow: scroll;
+  background: ${(props) => props.theme.colors.neutral10};
+  max-height: 300px;
+  & a {
+    color: ${(props) => props.theme.colors.primary};
+  }
+  ,
+  & > *:first-child {
+    margin-top: ${(props) => props.theme.spacing.none};
+  }
+  ,
+  & > *:last-child {
+    margin-bottom: ${(props) => props.theme.spacing.none};
   }
 `;
